backend/requirements.txt
django==5.2.8
djangorestframework>=3.15
python-dotenv>=1.0
django-cors-headers>=4.0
